Refactor module structure, restructure config command
